volume indicators isolation: https://github.com/bukosabino/ta/issues/129
