+++ ta/tests/__init__.py
@@ -1,7 +1,7 @@
-from ta.tests.momentum import (TestKAMAIndicator, TestMFIIndicator,
-                               TestRateOfChangeIndicator, TestRSIIndicator,
-                               TestStochasticOscillator, TestTSIIndicator,
-                               TestUltimateOscillator, TestWilliamsRIndicator)
+from ta.tests.momentum import (TestKAMAIndicator, TestRateOfChangeIndicator,
+                               TestRSIIndicator, TestStochasticOscillator,
+                               TestTSIIndicator, TestUltimateOscillator,
+                               TestWilliamsRIndicator)
 from ta.tests.trend import (TestADXIndicator, TestCCIIndicator,
                             TestMACDIndicator, TestPSARIndicator,
                             TestVortexIndicator)
@@ -9,5 +9,5 @@ from ta.tests.utils import TestGeneral
 from ta.tests.volatility import TestAverageTrueRange, TestBollingerBands
 from ta.tests.volume import (TestAccDistIndexIndicator,
                              TestEaseOfMovementIndicator,
-                             TestForceIndexIndicator,
+                             TestForceIndexIndicator, TestMFIIndicator,
                              TestOnBalanceVolumeIndicator)
